vat value type changed from Integer to bigDecimal

+++ src/main/java/cz/muni/fi/bapr/dao/VatDAO.java
@@ -2,11 +2,13 @@ package cz.muni.fi.bapr.dao;
 
 import cz.muni.fi.bapr.entity.Vat;
 
+import java.math.BigDecimal;
+
 /**
  * @author Andrej Kuročenko <andrej@kurochenko.net>
  */
 public interface VatDAO extends DAOTemplate<Vat> {
 
-    Vat findByVat(Integer vat);
+    Vat findByVat(BigDecimal vat);
 
 }
